requirements.txt
Flask==2.3.3
Werkzeug==2.3.8
PyMuPDF==1.24.2
openai==1.30.1
python-dotenv==1.0.1
transformers==4.41.2
google-generativeai==0.4.1
